+++ snapshot.gql
@@ -1,4 +1,4 @@
-### Type definitions saved at 2023-02-02T11:43:20.052Z ###
+### Type definitions saved at 2023-02-03T08:24:28.076Z ###
 
 enum RemoteFileFit {
   COVER
