+++ src/pages/comunicati-stampa.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import { graphql, Link } from 'gatsby';
+import { JobEntry } from '../components/Jobposition/JobEntry';
+import { Layout } from '../partials/Layout';
+
+const ComunicatiStampa = ({ data }: { data: Queries.AllJobsQuery }) => (
+  <Layout>
+    ComunicatiStampa
+  </Layout>
+);
+
+export default ComunicatiStampa;
